refactor: decouple tab-lifecycle.js from TabManager ctx
- Define TabLifecycleCtx JSDoc typedef for narrow interface
- Decouple findTabForTerminal: accepts tabs Map directly
- Decouple onTerminalCwdChanged: accepts tabs + activeTabId
- Document remaining functions with narrow interface type

Also fix broken context-menu.js import paths (from #41).

Closes #47
